# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CODO API
  description: Source-first legal information API for Côte d'Ivoire.
  version: 0.8.0
paths:
...
  /v1/me:
    get:
      tags:
      - user-account
      summary: Profile
      operationId: profile_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileOut'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - user-account
      summary: Delete Profile
      operationId: delete_profile_v1_me_delete
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
    patch:
      tags:
      - user-account
      summary: Update Profile
      operationId: update_profile_v1_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/accessibility:
    get:
      tags:
      - user-account
      summary: Accessibility
      operationId: accessibility_v1_me_accessibility_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessibilitySettingsOut'
      security:
      - HTTPBearer: []
    put:
      tags:
      - user-account
      summary: Update Accessibility
      operationId: update_accessibility_v1_me_accessibility_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessibilitySettingsIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessibilitySettingsOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/favorites:
    get:
      tags:
      - user-account
      summary: Favorites
      operationId: favorites_v1_me_favorites_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FavoriteOut'
                type: array
                title: Response Favorites V1 Me Favorites Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - user-account
      summary: Add Favorite
      operationId: add_favorite_v1_me_favorites_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/favorites/{favorite_id}:
    delete:
      tags:
      - user-account
      summary: Delete Favorite
      operationId: delete_favorite_v1_me_favorites__favorite_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: favorite_id
        in: path
        required: true
        schema:
          type: integer
          title: Favorite Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me/folders:
    get:
      tags:
      - user-account
      summary: Folders
      operationId: folders_v1_me_folders_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserFolderOut'
                type: array
                title: Response Folders V1 Me Folders Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - user-account
      summary: Create Folder
      operationId: create_folder_v1_me_folders_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFolderCreateIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFolderOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/folders/{folder_id}:
    delete:
      tags:
      - user-account
      summary: Delete Folder
      operationId: delete_folder_v1_me_folders__folder_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Folder Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me/folders/{folder_id}/items:
    get:
      tags:
      - user-account
      summary: Folder Items
      operationId: folder_items_v1_me_folders__folder_id__items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Folder Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserFolderItemOut'
                title: Response Folder Items V1 Me Folders  Folder Id  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - user-account
      summary: Add Folder Item
      operationId: add_folder_item_v1_me_folders__folder_id__items_post
      security:
      - HTTPBearer: []
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Folder Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFolderItemIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFolderItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me/folders/{folder_id}/items/{item_id}:
    delete:
      tags:
      - user-account
      summary: Delete Folder Item
      operationId: delete_folder_item_v1_me_folders__folder_id__items__item_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Folder Id
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me/alerts:
    get:
      tags:
      - user-account
      summary: Alerts
      operationId: alerts_v1_me_alerts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserAlertOut'
                type: array
                title: Response Alerts V1 Me Alerts Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - user-account
      summary: Create Alert
      operationId: create_alert_v1_me_alerts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAlertCreateIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAlertOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/alerts/{alert_id}:
    patch:
      tags:
      - user-account
      summary: Update Alert
      operationId: update_alert_v1_me_alerts__alert_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAlertUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAlertOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - user-account
      summary: Delete Alert
      operationId: delete_alert_v1_me_alerts__alert_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccessibilitySettingsIn:
      properties:
        text_scale:
          type: number
          maximum: 1.8
          minimum: 0.8
          title: Text Scale
          default: 1.0
        high_contrast:
          type: boolean
          title: High Contrast
          default: false
        reduce_motion:
          type: boolean
          title: Reduce Motion
          default: false
        offline_cache_enabled:
          type: boolean
          title: Offline Cache Enabled
          default: false
      type: object
      title: AccessibilitySettingsIn
    AccessibilitySettingsOut:
      properties:
        text_scale:
          type: number
          maximum: 1.8
          minimum: 0.8
          title: Text Scale
          default: 1.0
        high_contrast:
          type: boolean
          title: High Contrast
          default: false
        reduce_motion:
          type: boolean
          title: Reduce Motion
          default: false
        offline_cache_enabled:
          type: boolean
          title: Offline Cache Enabled
          default: false
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - updated_at
      title: AccessibilitySettingsOut
...
    FavoriteIn:
      properties:
        entity_type:
          type: string
          enum:
          - document
          - article
          - procedure
          - institution
          - case_law
          title: Entity Type
        entity_key:
          type: string
          maxLength: 240
          minLength: 1
          title: Entity Key
      type: object
      required:
      - entity_type
      - entity_key
      title: FavoriteIn
    FavoriteOut:
      properties:
        id:
          type: integer
          title: Id
        entity_type:
          type: string
          enum:
          - document
          - article
          - procedure
          - institution
          - case_law
          title: Entity Type
        entity_key:
          type: string
          title: Entity Key
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - entity_type
      - entity_key
      - created_at
      title: FavoriteOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserAlertCreateIn:
      properties:
        scope_type:
          type: string
          enum:
          - domain
          - document
          - article
          - topic
          title: Scope Type
        scope_key:
          type: string
          maxLength: 240
          minLength: 1
          title: Scope Key
      type: object
      required:
      - scope_type
      - scope_key
      title: UserAlertCreateIn
    UserAlertOut:
      properties:
        id:
          type: string
          title: Id
        scope_type:
          type: string
          enum:
          - domain
          - document
          - article
          - topic
          title: Scope Type
        scope_key:
          type: string
          title: Scope Key
        enabled:
          type: boolean
          title: Enabled
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - scope_type
      - scope_key
      - enabled
      - created_at
      title: UserAlertOut
    UserAlertUpdateIn:
      properties:
        enabled:
          type: boolean
          title: Enabled
      type: object
      required:
      - enabled
      title: UserAlertUpdateIn
    UserFolderCreateIn:
      properties:
        name:
          type: string
          maxLength: 160
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      title: UserFolderCreateIn
    UserFolderItemIn:
      properties:
        entity_type:
          type: string
          enum:
          - document
          - article
          - procedure
          - institution
          - case_law
          - search
          title: Entity Type
        entity_key:
          type: string
          maxLength: 240
          minLength: 1
          title: Entity Key
      type: object
      required:
      - entity_type
      - entity_key
      title: UserFolderItemIn
    UserFolderItemOut:
      properties:
        id:
          type: integer
          title: Id
        entity_type:
          type: string
          enum:
          - document
          - article
          - procedure
          - institution
          - case_law
          - search
          title: Entity Type
        entity_key:
          type: string
          title: Entity Key
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - entity_type
      - entity_key
      - created_at
      title: UserFolderItemOut
    UserFolderOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        item_count:
          type: integer
          title: Item Count
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - created_at
      - updated_at
      title: UserFolderOut
    UserProfileOut:
      properties:
        id:
          type: string
          title: Id
        auth_subject:
          type: string
          title: Auth Subject
        auth_issuer:
          type: string
          title: Auth Issuer
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        email_verified:
          type: boolean
          title: Email Verified
          default: false
        locale:
          type: string
          title: Locale
          default: fr-CI
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
      type: object
      required:
      - id
      - auth_subject
      - auth_issuer
      - created_at
      - updated_at
      title: UserProfileOut
    UserProfileUpdateIn:
      properties:
        display_name:
          anyOf:
          - type: string
            maxLength: 240
            minLength: 1
          - type: 'null'
          title: Display Name
        locale:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 2
          - type: 'null'
          title: Locale
      type: object
      title: UserProfileUpdateIn
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer